requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu126
torch==2.9.0+cu126
torchvision==0.24.0+cu126
pillow~=12.0.0
onnxruntime~=1.23.2
numpy~=2.2.6
opencv-python~=4.12.0.88
pyyaml~=6.0.3